requirements.txt
functions-framework==3.*
numpy
scikit-learn==1.3.2
joblib==1.3.2
google-cloud-storage
gcsfs
